BayesianNetworkRegression.Fit! — Method

Fit!(X::AbstractArray{T}, y::AbstractVector{U}, R; η=1.01,ζ=1.0,ι=1.0,aΔ=1.0,bΔ=1.0, 
     ν=10, nburn=30000, nsamples=20000, V=0, x_transform=true, suppress_timer=false, num_chains=2, seed=nothing, 
     in_seq=false, full_results=false, purge_burn=nothing) where {T,U}

Fit the Bayesian Network Regression model, generating nsamples Gibbs samples after nburn burn-in are discarded

Arguments

X: matrix, required, matrix of unweighted symmetric adjacency matrices to be used as predictors. Two options: 2D matrix with each row the upper triangle of the adjacency matrix associated with one sample 1D matrix with each row the adjacency matrix relating the nodes to one another

y: vector, required, vector of response variables

R: integer, required, the dimensionality of the latent variables u, a hyperparameter

η: float, default=1.01, hyperparameter used for sampling the 0 value of the πᵥ parameter, must be > 1

ζ: float, default=1.0, hyperparameter used for sampling θ

ι: float, default=1.0, hyperparameter used for sampling θ

aΔ: float, default=1.0, hyperparameter used for sampling Δ

bΔ: float, default=1.0, hyperparameter used for sampling Δ

ν: integer, default=10, hyperparameter used for sampling M, must be > R

nburn: integer, default=30000, number of burn-in samples to generate and discard

nsamples: integer, default=20000, number of Gibbs samples to generate after burn-in

V: integer, default=0, dimensionality of the adjacency matrices (number of nodes)

x_transform: boolean, default=true, set to false if X has been pre-transformed into one row per sample. Otherwise the X will be transformed automatically.

suppress_timer: boolean, default=false, set to true to suppress "progress meter" output

num_chains: integer, default=2, number of separate sampling chains to run (for checking convergence)

seed: integer, default=nothing, random seed used for repeatability

in_seq: boolean, default=false, set to true to run multiple chains in sequence. Otherwise multiple chains will be run in parallel. (for performance reasons should be set to false if only one chain is used)

full_results: boolean, default=false, set to true to return full post-burn-in chains for relevant variables

purge_burn: integer, default=nothing, if set must be less than the number of burn-in samples (and ideally burn-in is a multiple of this value). After how many burn-in samples to delete previous burn-in samples.

Returns

Either the entire state table with post-burn-in samples of relevant variables (ξ, γ, μ, τ²) or a Results object with the state table from the first chain and PSRF r-hat values for γ and ξ (depending on the value of full_results)